// apps/web-svelte/src/lib/api/generated/model/providerLibraryConnection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ProviderLibraryDescriptor } from './providerLibraryDescriptor';

export interface ProviderLibraryConnection {
  connectionId: string;
  connectionName: string;
  pluginId: string;
  libraries: ProviderLibraryDescriptor[];
  /** @nullable */
  error?: string | null;
}